Refactoring codesync into its own file
